pyproject.toml
[tool.poetry]
name = "parcellab"
version = "0.1.0"
description = ""
authors = ["Friday Godswill <faradayyg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.2"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
python-decouple = "^3.8"
pytest-django = "^4.5.2"
pycountry = "^22.3.5"
structlog = "^23.1.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.1.0"
ipdb = "^0.13.13"
pytest = "^7.3.2"
django-extensions = "^3.2.3"
pre-commit = "^3.3.3"
mypy = "^1.3.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "parcelLab.settings"
python_files = ["tests/*/test_*"]

[tool.ruff]
exclude = ["parcelLab/settings.py"]
